requirements.txt
# Phase 1 Requirements
kafka-python==2.0.2
pymysql==1.0.2
flask==2.3.2
sqlalchemy==2.0.19
pandas==2.0.3
numpy==1.24.3
confluent-kafka==2.1.1
requests==2.31.0
python-dotenv==1.0.0

# Phase 2 Requirements
pyspark==3.4.1
matplotlib==3.7.1
scipy==1.10.1
kubernetes==25.3.0
seaborn==0.12.2
jupyter==1.0.0